stream: move tests into tests/
